Switch to GTCEu/RFG Build Scripts
Deploy and configure Migration.zip so the build works again
