+++ .gitignore
@@ -1,32 +1,83 @@
-# eclipse
-bin
-*.launch
-.settings
+.nopublish
+
+### Windows ###
+
+thumbs.db
+*.db
+
+### Java ###
+
+*.class
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.war
+*.ear
+*.txt
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+### Eclipse ###
+
+*.pydevproject
 .metadata
-.classpath
+.gradle
+bin/
+tmp/
+*.tmp
+*.bak
+*.swp
+*~.nib
+local.properties
+.settings/
+.loadpath
+/eclipse
+
+# Eclipse Core
 .project
 
-# idea
-out
+# External tool builders
+.externalToolBuilders/
+
+# Locally stored "Eclipse launch configurations"
+*.launch
+
+# CDT-specific
+.cproject
+
+# JDT-specific (Eclipse Java Development Tools)
+.classpath
+
+# Java annotation processor (APT)
+.factorypath
+
+# PDT-specific
+.buildpath
+
+# sbteclipse plugin
+.target
+
+# TeXlipse plugin
+.texlipse
+
+### Intellij IDEA ###
+
+*.iml
 *.ipr
 *.iws
-*.iml
-.idea
 
-# gradle
-build
-.gradle
-
-# other
-eclipse
-run
-
-# Files from Forge MDK
-forge*changelog.txt
-/run/
+.idea/
+.idea_modules/
+/classes/
+/out/
 /build/
-/.gradle/
-/.idea/
-/.settings/
-/FactoryBuilderDiagram.xcf
-/examples/tests/
+
+# Linux
+*~
+
+run/
+
+logs/
