+++ src/main/java/eutros/multiblocktweaker/MultiblockTweaker.java
@@ -1,13 +1,14 @@
 package eutros.multiblocktweaker;
 
 import eutros.multiblocktweaker.client.PreviewHandler;
+import gregtech.GTInternalTags;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 
 @Mod(modid = MultiblockTweaker.MOD_ID,
         name = "Multiblock Tweaker",
-        version = "@GRADLE:VERSION@",
-        dependencies = "required-after:gregtech@[1.17.0.764,);")
+        version = MBTTags.VERSION,
+        dependencies = GTInternalTags.DEP_VERSION_STRING)
 public class MultiblockTweaker {
 
     public static final String MOD_ID = "multiblocktweaker";
